Change the function upgrading packages so that it changes the behaviour according with your distribution

+++ .bashrc
@@ -7,8 +7,8 @@ mkcd() {
 }
 
 update() {
-  echo -e "### APT ###"
-  update_apt
+  echo -e "### packages ###"
+  update_packages
   echo -e "\n### RBENV ###"
   update_rbenv
   echo -e "\n### GEM ###"
@@ -25,11 +25,31 @@ update() {
   update_rust
 }
 
-update_apt() {
-  sudo apt-get update -y
-  sudo apt-get upgrade -y
-  sudo apt-get autoclean -y
-  sudo apt-get autoremove -y
+update_packages() {
+  if [[ ! -f /etc/os-release ]]; then
+    echo "/etc/os-release does not exist."
+    return 1
+  fi
+
+  . /usr/lib/os-release
+  case $ID in
+    debian | ubuntu )
+      sudo apt-get update -y
+      sudo apt-get upgrade -y
+      sudo apt-get autoclean -y
+      sudo apt-get autoremove -y;;
+    arch )
+      sudo pacman -Syu
+      pacman -Qdtq | pacman -Rs -
+      paccache -ruk0;;
+    rhel ) echo "Red Hat";;
+    centos ) echo "CentOS";;
+    fedora ) echo "Fedora";;
+    opensuse ) echo "OpenSUSE";;
+    * )
+      echo "Unknown distribution"
+      return 1;;
+  esac
 }
 
 update_rbenv() {
